refactor(nms): Consolidate SubscriberContext
