+++ nms/app/components/lte/LteContext.tsx
@@ -11,118 +11,23 @@
  * limitations under the License.
  */
 import * as React from 'react';
-import InitSubscriberState, {
-  FetchSubscriberState,
-  getGatewaySubscriberMap,
-  setSubscriberState,
-} from '../../state/lte/SubscriberState';
-import LoadingFiller from '../LoadingFiller';
-import SubscriberContext from '../context/SubscriberContext';
 import {ApnContextProvider} from '../context/ApnContext';
 import {CbsdContextProvider} from '../context/CbsdContext';
 import {EnodebContextProvider} from '../context/EnodebContext';
-import {
-  FEG_LTE,
-  LTE,
-  NetworkId,
-  SubscriberId,
-} from '../../../shared/types/network';
+import {FEG_LTE, LTE, NetworkId} from '../../../shared/types/network';
 import {GatewayContextProvider} from '../context/GatewayContext';
 import {GatewayPoolsContextProvider} from '../context/GatewayPoolsContext';
 import {GatewayTierContextProvider} from '../context/GatewayTierContext';
 import {LteNetworkContextProvider} from '../context/LteNetworkContext';
 import {PolicyProvider} from '../context/PolicyContext';
+import {SubscriberContextProvider} from '../context/SubscriberContext';
 import {TraceContextProvider} from '../context/TraceContext';
-import {useEffect, useState} from 'react';
-import {useEnqueueSnackbar} from '../../hooks/useSnackbar';
-import type {
-  MutableSubscriber,
-  Subscriber,
-  SubscriberState,
-} from '../../../generated';
 
 type Props = {
   networkId: NetworkId;
   networkType: string;
   children: React.ReactNode;
 };
-
-export function SubscriberContextProvider(props: Props) {
-  const {networkId} = props;
-  const [subscriberMap, setSubscriberMap] = useState<
-    Record<string, Subscriber>
-  >({});
-  const [sessionState, setSessionState] = useState<
-    Record<string, SubscriberState>
-  >({});
-  const [subscriberMetrics, setSubscriberMetrics] = useState({});
-  const [isLoading, setIsLoading] = useState(true);
-  const [totalCount, setTotalCount] = useState(0);
-  const enqueueSnackbar = useEnqueueSnackbar();
-  useEffect(() => {
-    const fetchLteState = async () => {
-      if (networkId == null) {
-        return;
-      }
-      await InitSubscriberState({
-        networkId,
-        setSubscriberMap,
-        setSubscriberMetrics,
-        setSessionState,
-        setTotalCount,
-        enqueueSnackbar,
-      });
-      setIsLoading(false);
-    };
-    void fetchLteState();
-  }, [networkId, enqueueSnackbar]);
-
-  if (isLoading) {
-    return <LoadingFiller />;
-  }
-
-  return (
-    <SubscriberContext.Provider
-      value={{
-        forbiddenNetworkTypes: {},
-        state: subscriberMap,
-        metrics: subscriberMetrics,
-        sessionState: sessionState,
-        totalCount: totalCount,
-        gwSubscriberMap: getGatewaySubscriberMap(sessionState),
-        setState: (
-          key: SubscriberId,
-          value?: MutableSubscriber | Array<MutableSubscriber>,
-          newState?,
-        ) =>
-          setSubscriberState({
-            networkId,
-            subscriberMap,
-            setSubscriberMap,
-            setSessionState,
-            key,
-            value,
-            newState,
-          }),
-        refetchSessionState: (id?: SubscriberId) => {
-          void FetchSubscriberState({networkId, id}).then(sessions => {
-            if (sessions) {
-              setSessionState(currentSessionState =>
-                id
-                  ? {
-                      ...currentSessionState,
-                      ...sessions,
-                    }
-                  : sessions,
-              );
-            }
-          });
-        },
-      }}>
-      {props.children}
-    </SubscriberContext.Provider>
-  );
-}
 
 export function LteContextProvider(props: Props) {
   const {networkId, networkType} = props;
@@ -132,16 +37,16 @@ export function LteContextProvider(props: Props) {
   }
 
   return (
-    <LteNetworkContextProvider {...{networkId, networkType}}>
+    <LteNetworkContextProvider networkId={networkId}>
       <PolicyProvider networkId={networkId}>
         <ApnContextProvider networkId={networkId}>
-          <SubscriberContextProvider {...{networkId, networkType}}>
-            <GatewayTierContextProvider {...{networkId, networkType}}>
+          <SubscriberContextProvider networkId={networkId}>
+            <GatewayTierContextProvider networkId={networkId}>
               <EnodebContextProvider networkId={networkId}>
                 <GatewayContextProvider networkId={networkId}>
                   <GatewayPoolsContextProvider networkId={networkId}>
                     <TraceContextProvider networkId={networkId}>
-                      <CbsdContextProvider {...{networkId, networkType}}>
+                      <CbsdContextProvider networkId={networkId}>
                         {props.children}
                       </CbsdContextProvider>
                     </TraceContextProvider>
