+++ nms/app/views/subscriber/SubscriberUtils.tsx
@@ -19,7 +19,7 @@ import React, {useContext} from 'react';
 import ReactJson from 'react-json-view';
 import SubscriberContext from '../../components/context/SubscriberContext';
 import {MutableSubscriber, Subscriber} from '../../../generated';
-import {SubscriberRowType} from '../../state/lte/SubscriberState';
+import {SubscriberRowType} from '../../util/SubscriberState';
 import {isValidHex} from '../../util/strings';
 import {useNavigate} from 'react-router-dom';
 import type {
